PoS: move rewards and slashing fns into their mods
