+++ crates/proof_of_stake/src/lib.rs
@@ -28,10 +28,10 @@ use std::collections::{BTreeMap, BTreeSet, HashSet};
 
 pub use error::*;
 use namada_core::event::EmitEvents;
-use namada_core::tendermint::abci::types::{Misbehavior, MisbehaviorKind};
+use namada_core::tendermint::abci::types::Misbehavior;
 use namada_core::types::address::{Address, InternalAddress};
 use namada_core::types::dec::Dec;
-use namada_core::types::key::{common, tm_raw_hash_to_string};
+use namada_core::types::key::common;
 use namada_core::types::storage::BlockHeight;
 pub use namada_core::types::storage::{Epoch, Key, KeySeg};
 use namada_storage::collections::lazy_map::{self, Collectable, LazyMap};
@@ -2806,7 +2806,7 @@ where
     // because it potentially needs to be able to read validator state from
     // previous epoch and jailing validator removes the historical state
     if !votes.is_empty() {
-        log_block_rewards(
+        rewards::log_block_rewards(
             storage,
             votes.clone(),
             height,
@@ -2817,7 +2817,7 @@ where
 
     // Invariant: This has to be applied after
     // `copy_validator_sets_and_positions` and before `self.update_epoch`.
-    record_slashes_from_evidence(
+    slashing::record_slashes_from_evidence(
         storage,
         byzantine_validators,
         &pos_params,
@@ -2874,145 +2874,5 @@ where
         prune_liveness_data(storage, current_epoch)?;
     }
 
-    Ok(())
-}
-
-// Process the proposer and votes in the block to assign their PoS rewards.
-fn log_block_rewards<S>(
-    storage: &mut S,
-    votes: Vec<VoteInfo>,
-    height: BlockHeight,
-    current_epoch: Epoch,
-    new_epoch: bool,
-) -> namada_storage::Result<()>
-where
-    S: StorageWrite + StorageRead,
-{
-    // Read the block proposer of the previously committed block in storage
-    // (n-1 if we are in the process of finalizing n right now).
-    match storage::read_last_block_proposer_address(storage)? {
-        Some(proposer_address) => {
-            tracing::debug!("Found last block proposer: {proposer_address}");
-            rewards::log_block_rewards(
-                storage,
-                if new_epoch {
-                    current_epoch.prev()
-                } else {
-                    current_epoch
-                },
-                &proposer_address,
-                votes,
-            )?;
-        }
-        None => {
-            if height > BlockHeight::default().next_height() {
-                tracing::error!(
-                    "Can't find the last block proposer at height {height}"
-                );
-            } else {
-                tracing::debug!("No last block proposer at height {height}");
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Apply PoS slashes from the evidence
-fn record_slashes_from_evidence<S>(
-    storage: &mut S,
-    byzantine_validators: Vec<Misbehavior>,
-    pos_params: &PosParams,
-    current_epoch: Epoch,
-    validator_set_update_epoch: Epoch,
-) -> namada_storage::Result<()>
-where
-    S: StorageWrite + StorageRead,
-{
-    if !byzantine_validators.is_empty() {
-        let pred_epochs = storage.get_pred_epochs()?;
-        for evidence in byzantine_validators {
-            // dbg!(&evidence);
-            tracing::info!("Processing evidence {evidence:?}.");
-            let evidence_height = match u64::try_from(evidence.height) {
-                Ok(height) => height,
-                Err(err) => {
-                    tracing::error!("Unexpected evidence block height {}", err);
-                    continue;
-                }
-            };
-            let evidence_epoch =
-                match pred_epochs.get_epoch(BlockHeight(evidence_height)) {
-                    Some(epoch) => epoch,
-                    None => {
-                        tracing::error!(
-                            "Couldn't find epoch for evidence block height {}",
-                            evidence_height
-                        );
-                        continue;
-                    }
-                };
-            // Disregard evidences that should have already been processed
-            // at this time
-            if evidence_epoch + pos_params.slash_processing_epoch_offset()
-                - pos_params.cubic_slashing_window_length
-                <= current_epoch
-            {
-                tracing::info!(
-                    "Skipping outdated evidence from epoch {evidence_epoch}"
-                );
-                continue;
-            }
-            let slash_type = match evidence.kind {
-                MisbehaviorKind::DuplicateVote => {
-                    types::SlashType::DuplicateVote
-                }
-                MisbehaviorKind::LightClientAttack => {
-                    types::SlashType::LightClientAttack
-                }
-                MisbehaviorKind::Unknown => {
-                    tracing::error!("Unknown evidence: {:#?}", evidence);
-                    continue;
-                }
-            };
-            let validator_raw_hash =
-                tm_raw_hash_to_string(evidence.validator.address);
-            let validator = match storage::find_validator_by_raw_hash(
-                storage,
-                &validator_raw_hash,
-            )? {
-                Some(validator) => validator,
-                None => {
-                    tracing::error!(
-                        "Cannot find validator's address from raw hash {}",
-                        validator_raw_hash
-                    );
-                    continue;
-                }
-            };
-            // Check if we're gonna switch to a new epoch after a delay
-            tracing::info!(
-                "Slashing {} for {} in epoch {}, block height {} (current \
-                 epoch = {}, validator set update epoch = \
-                 {validator_set_update_epoch})",
-                validator,
-                slash_type,
-                evidence_epoch,
-                evidence_height,
-                current_epoch
-            );
-            if let Err(err) = slashing::slash(
-                storage,
-                pos_params,
-                current_epoch,
-                evidence_epoch,
-                evidence_height,
-                slash_type,
-                &validator,
-                validator_set_update_epoch,
-            ) {
-                tracing::error!("Error in slashing: {}", err);
-            }
-        }
-    }
     Ok(())
 }
